gen_ai/agents/pyproject.toml
[project]
name = "itinerary-planner"
version = "0.1.0"
description = "Agentic itinerary planner using LangGraph and OpenRouter"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langgraph-cli[inmem]",
    "openai>=1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0",
    "loguru>=0.7.0",
    "ddgs>=9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
itinerary-planner = "itinerary_planner.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/itinerary_planner"]

[dependency-groups]
dev = []
